feat: moved auth cookie responsibility to response/endpoint
